AlphaBeta 알고리즘 수정 (Best Move를 찾을 수 있도록)

+++ AI_Connect4/Random.cpp
@@ -4,9 +4,9 @@
 
 uint32_t XOR128::seed = 1;
 
-float XOR128::Random()
+double XOR128::Random()
 {
-	return (float) xor128() / 0xffffffff;
+	return (double) xor128() / (0xffffffff);
 }
 
 uint32_t XOR128::xor128()
